feat(ztnet webhooks): Base working implementation
TODO: Right now, the code is generic (can be catch-all for any webhook that doesn't have a payload that Discord understands) but the `/webhook` REST API POST endpoint is hardcoded (by env vars) to the ZTNET controller channel. If we fix, there needs to be a way to get the Discord webhook URL by making the Discord webhook ID and token of the URL part of the `/webhook` path/string.